% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot_extract_sparse.R
\name{autoplot.extract_sparse}
\alias{autoplot.extract_sparse}
\title{Plots the output of extract_sparse function.}
\usage{
\method{autoplot}{extract_sparse}(object, ...)
}
\arguments{
\item{object}{The output of the function 'extract_sparse'}

\item{...}{Other arguments currently ignored}
}
\value{
A ggplot object.
}
\description{
Plots the two lines fitted to the unique, sorted log degrees of the graph.
}
\examples{
library(igraph)
gr <- sample_pa(10000, power = 1.2, directed = FALSE)
sparse <- extract_sparse(gr)
autoplot(sparse)

}
