% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot_separate_dense_sparse.R
\name{autoplot.separate_dense_and_sparse}
\alias{autoplot.separate_dense_and_sparse}
\title{Plots the output of separate_dense_and_sparse.}
\usage{
\method{autoplot}{separate_dense_and_sparse}(object, ...)
}
\arguments{
\item{object}{The output of the function 'separate_dense_and_sparse.'}

\item{...}{Other arguments currently ignored}
}
\value{
A ggplot object.
}
\description{
Plots the three lines fitted to the unique, sorted log degrees of the graph.
}
\examples{
library(igraph)
W <- matrix(0.1, nrow = 100, ncol = 100)
wts <- c(0.5, 0.3, 0.2)
ns <- 200
nd <- 100
p <- 0.5
gr <- sample_mixed_graph(W, wts, nd, ns, p = 0.1, option = 2)
out <- separate_dense_and_sparse(gr)
separate <- separate_dense_and_sparse(gr)
autoplot(separate)

}
