% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_exp_matrix}
\alias{create_exp_matrix}
\title{Creates an nxn exponential matrix}
\usage{
create_exp_matrix(nrow, scalar)
}
\arguments{
\item{nrow}{The dimension of the matrix}

\item{scalar}{The scalar in exp(-(i+j)/scalar)}
}
\value{
An nxn matrix
}
\description{
Creates an nxn matrix where the (i,j)th entry is exp(-(i+j)/scalar)
}
\examples{
W <- create_exp_matrix(100, 100)

}
