% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphon_generation.R
\name{empirical_graphon}
\alias{empirical_graphon}
\title{Computes empirical graphon from graph}
\usage{
empirical_graphon(gr, n = NULL)
}
\arguments{
\item{gr}{A graph}

\item{n}{Dimension of the graphon matrix}
}
\value{
The empirical graphon
}
\description{
Computes empirical graphon given a graph
}
\examples{
library(igraph)
gr <- sample_gnp(1000, p=0.2)
emp <-  empirical_graphon(gr, n = 100)

}
