% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphon_generation.R
\name{generate_star_union}
\alias{generate_star_union}
\title{Generates a sparse graph of star graphs}
\usage{
generate_star_union(wts, n)
}
\arguments{
\item{wts}{The proportion of the hub degrees}

\item{n}{The number of nodes in the new graph}
}
\value{
A disjoint union of star graphs
}
\description{
Generates a union of star graphs given the weights
}
\examples{
library(igraph)
wts <- c(0.5, 0.3, 0.2)
gr <- generate_star_union(wts, n = 100)
gr

}
