% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphonmix-package.R
\docType{package}
\name{graphonmix-package}
\alias{graphonmix}
\alias{graphonmix-package}
\title{graphonmix: Generates Dense and Sparse Graphs using Graphon Extensions}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Generates dense or sparse graphs using graphon mixtures and graphettes. Graphon mixtures uses two graphons U and W to generate graphs. Sparse graphs are generated in this case using the inverse line graph (root) operation. Graphettes have 3 components, the graphon W, a sparsifying sequence and a graph edit function. Both techniques can generate dense or sparse graphs. Kandanaarachchi and Ong (2026) \doi{10.48550/arXiv.2505.13864}, Wijesinghe et al (2026) \doi{10.48550/arXiv.2602.23566}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://sevvandi.github.io/graphonmix/}
}

}
\author{
\strong{Maintainer}: Sevvandi Kandanaarachchi \email{sevvandik@gmail.com} (\href{https://orcid.org/0000-0002-0337-0395}{ORCID})

}
\keyword{internal}
