% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hill_estimator.R
\name{hill_estimator}
\alias{hill_estimator}
\title{Computes the Hill estimator or the power of the degree distribution}
\usage{
hill_estimator(gr, min_deg = 2)
}
\arguments{
\item{gr}{A graph}

\item{min_deg}{The minimum degree to consider. Default set to 5}
}
\value{
The Hill estimate
}
\description{
Given a graph, computes the Hill estimator or the power of the
degree distribution.
}
\examples{
library(igraph)
gr <- sample_pa(10000,  directed = FALSE)
hill_estimator(gr)


}
