% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphon_generation.R
\name{line_graphon}
\alias{line_graphon}
\title{Creates a line graphon from a sequence of probabilities}
\usage{
line_graphon(probs)
}
\arguments{
\item{probs}{The list of probabilities starting from the largest.}
}
\value{
The line graphon
}
\description{
Creates a line graphon, which is a disjoint clique graphon from a
sequence of probability values
}
\examples{
library(ggplot2)
wts <- c(0.5, 0.3, 0.2)
U <- line_graphon(wts)
plot_graphon(U)

}
