% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_graphon}
\alias{plot_graphon}
\title{Plots graphon}
\usage{
plot_graphon(W, cols = c("white", "black"))
}
\arguments{
\item{W}{A graphon given by an nxn matrix}

\item{cols}{Colors, by default white and black}
}
\value{
A ggplot object.
}
\description{
Plots graphon
}
\examples{
W <- create_exp_matrix(100, 100)
plot_graphon(W)

}
