% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ring_graphon}
\alias{ring_graphon}
\title{Creates a ring graphon}
\usage{
ring_graphon(n, alpha = 0.1)
}
\arguments{
\item{n}{The dimension of the matrix}

\item{alpha}{A scaling factor used in the graphon. Default set to 0.1}
}
\value{
An nxn matrix
}
\description{
Creates an graphon which can generate ring graphs.
}
\examples{
# example code
W <- ring_graphon(100)
plot_graphon(W)
}
