% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sbm_graphon}
\alias{sbm_graphon}
\title{Creates a Stochastic Block Model graphon}
\usage{
sbm_graphon(mat, n)
}
\arguments{
\item{mat}{The matrix representing the SBM}

\item{n}{The dimension of the matrix}
}
\value{
An nxn matrix
}
\description{
Creates an graphon representing a Stochastic Block Model (SBM).
}
\examples{
# example code
mat <- matrix(c(0.9, 0.01, 0.02,
0.01, 0.8, 0.03,
0.02, 0.03, 0.7), nrow = 3, byrow = TRUE)
W <- sbm_graphon(mat, 100)
plot_graphon(W)
}
