% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphettes.R
\name{star_f1}
\alias{star_f1}
\title{Adds stars to nodes}
\usage{
star_f1(xvals, t)
}
\arguments{
\item{xvals}{Values between 0 and 1 from the graphon}

\item{t}{A parameter indicating the strength of hubs}
}
\value{
a vector of values
}
\description{
Adds stars to nodes
}
\examples{
# example code
st <- star_f1(runif(10), 3)
st
}
