% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphonmix-package.R
\docType{package}
\name{graphonmix-package}
\alias{graphonmix}
\alias{graphonmix-package}
\title{graphonmix: Generates Mixture Graphs from Dense and Sparse Graphons}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Generates (U,W) mixture graphs where U is a line graph graphon and W is a dense graphon. Graphons are graph limits and graphon U can be written as sequence of positive numbers adding to 1. Graphs are sampled from U and W and joined randomly to obtain the mixture graph. Given a mixture graph, U can be inferred. Kandanaarachchi and Ong (2025) \doi{10.48550/arXiv.2505.13864}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://sevvandi.github.io/graphonmix/}
}

}
\author{
\strong{Maintainer}: Sevvandi Kandanaarachchi \email{sevvandik@gmail.com} (\href{https://orcid.org/0000-0002-0337-0395}{ORCID})

}
\keyword{internal}
